[dataset] Put the right stuff in the cache.
Unlike the previous commit, this didn't usually break anything.

The added test will break, but the only way to get into trouble was to use
both a dataset that resulted from link_to_here with override_previous that
happened to get a cache and then also use the original dataset. Very
unlikely to have given anyone problems in practice.

The fix does make the cache more effective, since it actually contains what
it's supposed to. (Not a big difference.)