build() can take unambiguous options/datasets/jobs as kws.
so instead of
	urd.build(
		'example',
		options=dict(foo=bar),
		datasets=dict(parent=ds),
	)

you can do
	urd.build('example', foo=bar, parent=ds)

Works on both urd.build and subjobs.build.

If you have both an option foo and a dataset foo you will have to
specify that as options=dict(foo=...), but other options can still
be passed as kws if you want. The same thing applies if you have an
option called datasets, or some other collision.